Keep migrator command and CLI in sync
This PR also removes the ability to use the Polaris CLI as a standalone tool. It is now only used as a plugin with Shopify CLI.
