+++ polaris-cli/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@shopify/polaris-cli",
+  "name": "@shopify/plugin-polaris",
   "version": "0.0.0",
   "description": "Commands for building Shopify Apps with Polaris",
   "license": "SEE LICENSE IN LICENSE.md",
@@ -13,26 +13,20 @@
     "access": "public",
     "@shopify:registry": "https://registry.npmjs.org"
   },
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "bin": "./bin/run",
+  "type": "module",
   "files": [
-    "/bin",
-    "/dist"
+    "dist"
   ],
   "scripts": {
     "build": "tsc -b",
-    "dev": "tsc -b -w",
     "lint": "TIMING=1 eslint --cache .",
     "clean": "rm -rf .turbo node_modules dist"
   },
   "dependencies": {
     "@oclif/core": "^1.13.10",
-    "@oclif/plugin-help": "^5.0.0",
     "@shopify/polaris-migrator": "^0.0.0"
   },
   "devDependencies": {
-    "ts-node": "^10.2.1",
     "typescript": "^4.7.4"
   },
   "engine-strict": true,
@@ -45,11 +39,7 @@
     "win32"
   ],
   "oclif": {
-    "bin": "polaris-cli",
     "commands": "dist/commands",
-    "plugins": [
-      "@oclif/plugin-help"
-    ],
     "topicSeparator": " ",
     "topics": {
       "polaris": {
