+++ polaris-migrator/src/index.ts
@@ -1,1 +1,3 @@
-export {run} from './cli';
+export {cli, cliInfo} from './cli';
+export {run} from './run';
+export type {Flags} from './run';
